+++ src/VaultSharp/V1/Commons/BackendConfig.cs
@@ -1,4 +1,5 @@
-﻿using System.Text.Json.Serialization;
+﻿using System.Collections.Generic;
+using System.Text.Json.Serialization;
 
 namespace VaultSharp.Core
 {
@@ -35,5 +36,30 @@ namespace VaultSharp.Core
         /// </value>
         [JsonPropertyName("max_lease_ttl")]
         public int MaximumLeaseTtl { get; set; }
+        
+        [JsonPropertyName("description")]
+        public string Description { get; set; }
+
+        [JsonPropertyName("audit_non_hmac_request_keys")]
+        public List<string> AuditNonHmacRequestKeys { get; set; }
+
+        [JsonPropertyName("audit_non_hmac_response_keys")]
+        public List<string> AuditNonHmacResponseKeys { get; set; }
+
+        [JsonPropertyName("listing_visibility")]
+        public BackendListingVisibility ListingVisibility { get; set; } = BackendListingVisibility.hidden;
+
+        [JsonPropertyName("passthrough_request_headers")]
+        public List<string> PassthroughRequestHeaders { get; set; }
+
+        [JsonPropertyName("allowed_response_headers")]
+        public List<string> AllowedResponseHeaders { get; set; }
+
+        [JsonPropertyName("allowed_managed_keys")]
+        public List<string> AllowedManagedKeys { get; set; }
+
+        [JsonPropertyName("plugin_version")]
+        [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)] // CRITICAL: Very very important
+        public string PluginVersion { get; set; }
     }
 }